pyproject.toml
[project]
name = "receita-federal-do-brasil-dados-publicos-cnpj"
version = "0.1.0"
description = "ETL para download e processamento dos dados públicos da Receita Federal - CNPJ"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pandas>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pathlib",
    "tenacity>=8.0.0",
    "rich>=13.0.0"
]
